myApiJWT/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
pydantic